# Simplfied correctness computing