This patch computes correctness with 2 passes by counting misplaced letters with an array instead of searching the word for each letter. In other words, O(N^2) → O(N * 2). I don't know exactly much faster this algorithm is, but I can confirm with testing it is not slower than the old code.