@jonhoo 2022-03-11
```
$ hyperfine -m 2 -n old './roget-main --implementation popular' -n new './roget-this --implementation popular'
Benchmark 1: old
  Time (mean ± σ):     399.0 ms ±   3.4 ms    [User: 396.6 ms, System: 2.2 ms]
  Range (min … max):   395.6 ms … 404.1 ms    7 runs

Benchmark 2: new
  Time (mean ± σ):     401.2 ms ±   1.6 ms    [User: 398.8 ms, System: 2.2 ms]
  Range (min … max):   400.1 ms … 404.7 ms    7 runs

Summary
  'old' ran
    1.01 ± 0.01 times faster than 'new'
$ hyperfine -m 2 -n old './roget-main --implementation cutoff' -n new './roget-this --implementation cutoff'
Benchmark 1: old
  Time (mean ± σ):      3.615 s ±  0.004 s    [User: 3.613 s, System: 0.000 s]
  Range (min … max):    3.611 s …  3.618 s    2 runs

Benchmark 2: new
  Time (mean ± σ):      3.404 s ±  0.002 s    [User: 3.401 s, System: 0.000 s]
  Range (min … max):    3.403 s …  3.405 s    2 runs

Summary
  'new' ran
    1.06 ± 0.00 times faster than 'old'
```

Interestingly enough, this only gives a very marginal speedup (~6%). Still happy to merge it as I think it reads nicer anyway, but interesting that it's so marginal.