@ironhaven 2022-03-11
I guess the lesson is that changing O(N^2) to O(N) does not get you a lot when N = 5.