@jonhoo 2022-03-11
Hehe, yes, exactly. Or that the compiler is _really_ good at optimizing the code somehow.